更新

+++ modules/QuoteServer/src/QuoteServer.Domain/Instrument/Ins.cs
@@ -38,11 +38,11 @@ namespace QuoteServer.Domain.Entities
         /// <summary>
         /// 多方币种
         /// </summary>
-        public long Long { get; set; }
+        public int Long { get; set; }
         /// <summary>
         /// 空方币种
         /// </summary>
-        public long Short { get; set; }
+        public int Short { get; set; }
 
         /// <summary>
         /// 价格转换系数   如接入BTCUSD行情，设置0.2 即缩小5倍
